refactor: centralize formatting functions
